# backend/swaggerApi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "digrowfa",
    "description": "Digrowfa API in NodeJs using Swagger",
    "contact": {
      "email": "abdulmanafp1966@gmail.com"
    },
    "license": {
      "name": "DIGROWFA",
      "url": "https://www.digrowfa.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Environment"
    },
    {
      "url": "http://localhost:3000",
      "description": "Demo Environment"
    }
  ],
...
  "paths": {
    "/api/course": {
      "get": {
        "tags": [
          "Course"
        ],
        "summary": "GET THE TOTAL COURSES",
        "description": "All courses",
        "responses": {
          "200": {
            "description": "Status True"
          },
          "401": {
            "description": "Unauthorized User"
          }
        }
      },
      "post": {
        "tags": [
          "Course"
        ],
        "summary": "ADDS A NEW COURSE",
        "produces": "application/json",
        "consumes": "multipart/form-data",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "my Course"
                  },
                  "description": {
                    "type": "string",
                    "example": "this a course about MERN stack"
                  },
                  "category_id": {
                    "type": "objectId",
                    "example": "62f4d3e954dcb4be011ed0c7"
                  },
                  "instructor_id": {
                    "type": "objectId",
                    "example": "62e3618facbb5c77dc0341fa"
                  },
                  "duration": {
                    "type": "string",
                    "example": "3 months"
                  },
                  "language_id": {
                    "type": "objectId",
                    "example": "62f4adfe78e329ec4807b99e"
                  },
                  "image": {
                    "type": "file",
                    "example": "62e765e1fea9c14e2840e39c"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verified SuccessFully"
          },
          "401": {
            "description": "Unauthorized access to API"
          }
        }
      }
    },
...
    "/api/category": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "get all categories",
        "description": "for testing the swagger doc",
        "responses": {
          "200": {
            "description": "Status True"
          },
          "401": {
            "description": "Unauthorized User"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "get all categories",
        "description": "for testing the swagger doc",
        "responses": {
          "200": {
            "description": "Status True"
          },
          "401": {
            "description": "Unauthorized User"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Category A"
                  },
                  "description": {
                    "type": "string",
                    "example": ""
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/language": {
      "get": {
        "parameters": [
          {
            "name": "x-client-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "properties": {
                "x-client-id": {
                  "type": "string",
                  "example": "XXXXXXXXXXXXXXXXX"
                }
              }
            }
          }
        ],
        "tags": [
          "Language"
        ],
        "summary": "get all languages",
        "description": "for testing the swagger doc",
        "responses": {
          "200": {
            "description": "Status True"
          },
          "401": {
            "description": "Unauthorized User"
          }
        }
      },
      "post": {
        "tags": [
          "Language"
        ],
        "summary": "get all Languages",
        "description": "for testing the swagger doc",
        "responses": {
          "200": {
            "description": "Status True"
          },
          "401": {
            "description": "Unauthorized User"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Hindi"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}